GAE/requirements.txt
diffusers==0.21.4
h5py==3.8.0
imageio==2.27.0
matplotlib==3.5.2
numpy==1.23.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
Pillow==9.5.0
Pillow==10.1.0
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.9.1
sewar==0.4.5
skimage==0.0
spectral==0.23.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
